Settings Beginning 7/16
The initial steps of server-specific settings begins.
A folder within the config folder exists called "settings." A .ini per server is created based on the default settings found hardcoded or in example_server.ini. The server admins can then proceed to configure the settings.
The predefined settings are basic and only slightly recommended.